requirements.txt
pillow
requests
pygame
aiohttp
async-timeout
websockets==13.1
yarl
pytest-asyncio
